refact: move functions for signing target file to apis, remove developer_tool
